services/analytics-service/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
asyncpg
sqlalchemy
sqlmodel
psycopg2-binary